+++ noxfile.py
@@ -1,8 +1,3 @@
 """nox-poetry configuration file."""
 
-from calcipy.dev.noxfile import build_check, build_dist, coverage, pin_dev_dependencies, tests  # noqa: F401
-
-# Ensure that non-calcipy dev-dependencies are available in Nox environments
-pin_dev_dependencies([
-    'pytest-shell-utilities>=1.7.0"',
-])
+from calcipy.noxfile import build_check, build_dist, tests  # noqa: F401